Make generic encode_to_datagrams
